refactor: align scaffolded data layer with schema updates across Core, Indexing, Rule, Permissions, ServiceApp, Test, and User contexts

+++ Infrastructure/OrbitHub.Data/ServiceAppManagement/ServiceAppAuthentication.cs
@@ -21,7 +21,7 @@ public partial class ServiceAppAuthentication
 	/// Unique identifier for this ServiceAppAuthentications record.
 	/// </summary>
 	[Column("Id", IsPrimaryKey = true, IsIdentity = true, SkipOnInsert = true, SkipOnUpdate = true)]
-	public long Id { get; set; } // bigint
+	public int Id { get; set; } // int
 
 	/// <summary>
 	/// Public identifier used by the UI and external systems (GUID).
